+++ CHANGELOG.rst
@@ -3,6 +3,7 @@ Change Log
 
 0.0.26
 -------
+- Improve the calculation method of industry PE to enhance accuracy[`#92 <https://github.com/defeat-beta/defeatbeta-api/issues/92>`_]
 
 0.0.25
 -------